+++ index.js
@@ -1,11 +1,7 @@
-const esri = require('./providers/esri')
-const bing = require('./providers/bing')
-const openstreetmap = require('./providers/openstreetmap')
-const toporama = require('./providers/toporama')
-
 module.exports = {
-  esri: esri,
-  bing: bing,
-  openstreetmap: openstreetmap,
-  toporama: toporama
+  esri: require('./providers/esri'),
+  bing: require('./providers/bing'),
+  openstreetmap: require('./providers/openstreetmap'),
+  toporama: require('./providers/toporama'),
+  digitalglobe: require('./providers/digitalglobe')
 }