+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 
 # Changelog
 
+## 1.2.0 - 2017-05-09
+
+- Add DigitalGlobe Imagery provider
+
 ## 1.1.0 - 2017-04-28
 
 - Add OpenStreetMap hillshade provider